requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Faker==37.4.0
flasgger==0.9.7.1
Flask==3.1.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.5
fpdf==1.7.2
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mistune==3.1.3
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prometheus_client==0.22.1
prometheus-flask-exporter==0.23.2
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
